+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### 2022
 
+- v1.6.0 (2022-09-27)
+  - change function `Bytes` to `Byte` and change its first param from `[]byte` to `string`
+  - this function now save to validate only ASCII characters
+
 - v1.4.0 - v1.5.1 (2022-09-26)
   - add proper benchmark
   - seperate validate function to `Bytes` and `String`